// web/src/proto/generated/AppHostMsg.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ApiError } from "./ApiError";
import type { Event } from "./Event";
import type { Response } from "./Response";

/**
 * Host -> client.
 */
export type AppHostMsg = { "t": "res", id: number, ok: boolean, data?: Response, error?: ApiError, } | { "t": "event", event: Event, };